Move repos state to context
