+++ src/App.js
@@ -24,28 +24,7 @@ const App = () => {
   const [showClearUsers, setShowClearUsers] = useState(false);
   const [alert, setAlert] = useState({ msg: '', type: '' });
 
-  const getUser = async (username) => {
-    setLoading(true);
-    const res = await axios.get(
-      `https://api.github.com/users/${username}?&client_id=${process.env.REACT_APP_GITHUB_CLIENT_ID}&client_secret=${process.env.REACT_APP_GITHUB_CLIENT_SECRET}}`
-    );
-    setUser(res.data);
-    setLoading(false);
-    setShowClearUsers(true);
-  };
-
-  const getUserRepos = async (username) => {
-    setLoading(true);
-    const res = await axios.get(
-      `https://api.github.com/users/${username}/repos?per_page=5&sort=created:asc&client_id=${process.env.REACT_APP_GITHUB_CLIENT_ID}&client_secret=${process.env.REACT_APP_GITHUB_CLIENT_SECRET}}`
-    );
-    setRepos(res.data);
-    setLoading(false);
-    setShowClearUsers(true);
-  };
-
- 
-
+  
   const showAlert = (msg, type) => {
     setAlert({
       msg,
@@ -73,9 +52,7 @@ const App = () => {
                 path="/"
                 render={(props) => (
                   <>
-                    <Search
-                      setAlert={showAlert}
-                    />
+                    <Search setAlert={showAlert} />
                     <Users />
                   </>
                 )}
@@ -86,11 +63,6 @@ const App = () => {
                 render={(props) => (
                   <User
                     {...props}
-                    getUser={getUser}
-                    getUserRepos={getUserRepos}
-                    user={user}
-                    repos={repos}
-                    loading={loading}
                   />
                 )}
               />
