+++ src/context/github/GithubState.js
@@ -45,8 +45,24 @@ const GithubState = (props) => {
   };
 
   // Get user
+  const getUser = async (username) => {
+    setLoading();
+    const res = await axios.get(
+      `https://api.github.com/users/${username}?&client_id=${process.env.REACT_APP_GITHUB_CLIENT_ID}&client_secret=${process.env.REACT_APP_GITHUB_CLIENT_SECRET}}`
+    );
+    dispatch({ type: GET_USER, payload: res.data });
+    clearUsers();
+  };
 
   // Get repos
+  const getUserRepos = async (username) => {
+    setLoading();
+    const res = await axios.get(
+      `https://api.github.com/users/${username}/repos?per_page=5&sort=created:asc&client_id=${process.env.REACT_APP_GITHUB_CLIENT_ID}&client_secret=${process.env.REACT_APP_GITHUB_CLIENT_SECRET}}`
+    );
+    dispatch({ type: GET_REPOS, payload: res.data });
+    clearUsers();
+  };
 
   // Clear users
   const clearUsers = () => dispatch({ type: CLEAR_USERS });
@@ -62,7 +78,9 @@ const GithubState = (props) => {
         repos: state.repos,
         loading: state.loading,
         searchUsers,
-        clearUsers
+        getUser,
+        clearUsers,
+        getUserRepos,
       }}
     >
       {props.children}
