+++ src/context/github/githubReducer.js
@@ -17,10 +17,24 @@ export default (state, action) => {
         loading: false,
       };
 
+    case GET_USER:
+      return {
+        ...state,
+        user: action.payload,
+        loading: false,
+      };
+
     case CLEAR_USERS:
       return {
         ...state,
         users: [],
+        loading: false,
+      };
+
+    case GET_REPOS:
+      return {
+        ...state,
+        repos: action.payload,
         loading: false,
       };
